move grid from scripts to grid
