+++ lizrd/grid/grid_entrypoint.sh
