+++ lizrd/grid/tests/test_yaml_handling.py
@@ -9,7 +9,7 @@ class TestYaml(GeneralTestCase):
         """
         Test if the yaml config with multiple configs is loaded correctly
         """
-        path_to_yaml = "lizrd/scripts/tests/configs/multiple_yamls.yaml"
+        path_to_yaml = "lizrd/grid/tests/configs/multiple_yamls.yaml"
         configs = yaml.safe_load_all(open(path_to_yaml))
         configs = list(configs)
         for config in configs:
@@ -23,11 +23,11 @@ class TestYaml(GeneralTestCase):
         """
         Test if the yaml config with  inheritance is loaded correctly
         """
-        path_to_yaml = "lizrd/scripts/tests/configs/inheritance_yaml.yaml"
+        path_to_yaml = "lizrd/grid/tests/configs/inheritance_yaml.yaml"
         configs, paths_to_all_configs = load_with_inheritance(path_to_yaml)
         assert paths_to_all_configs == [
-            "lizrd/scripts/tests/configs/inheritance_yaml.yaml",
-            "lizrd/scripts/tests/configs/inheritance_yaml2.yaml",
+            "lizrd/grid/tests/configs/inheritance_yaml.yaml",
+            "lizrd/grid/tests/configs/inheritance_yaml2.yaml",
         ]
         configs = list(configs)
         for config in configs:
