+++ lizrd/test/test_grid.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import patch
 
-from lizrd.scripts.grid import create_subprocess_args
+from lizrd.grid.grid import create_subprocess_args
 from lizrd.grid.infrastructure import (
     IdeasBackend,
     EntropyBackend,
@@ -95,7 +95,7 @@ class TestGrid(unittest.TestCase):
                 "--time=5-05:00:00",
                 "--mem=32G",
                 None,
-                "lizrd/scripts/grid_entrypoint.sh",
+                "lizrd/grid/grid_entrypoint.sh",
                 "singularity",
                 "run",
                 "--env",
@@ -194,7 +194,7 @@ class TestGrid(unittest.TestCase):
                 "--mem=125G",
                 "--job-name=granular_4_mini",
                 "--time=0-05:00:00",
-                "lizrd/scripts/grid_entrypoint.sh",
+                "lizrd/grid/grid_entrypoint.sh",
                 "singularity",
                 "run",
                 "--env",
@@ -290,7 +290,7 @@ class TestGrid(unittest.TestCase):
                 "--mem=125G",
                 "--job-name=granular_4_mini",
                 "--time=0-05:00:00",
-                "lizrd/scripts/grid_entrypoint.sh",
+                "lizrd/grid/grid_entrypoint.sh",
                 "singularity",
                 "run",
                 "--env",
@@ -405,7 +405,7 @@ class TestGrid(unittest.TestCase):
                 "--account=plgsubslearnath-gpu-a100",
                 "--job-name=lr_grid",
                 "--time=40:00:00",
-                "lizrd/scripts/grid_entrypoint.sh",
+                "lizrd/grid/grid_entrypoint.sh",
                 "singularity",
                 "run",
                 "--bind=/net:/net",
@@ -466,7 +466,7 @@ class TestGrid(unittest.TestCase):
                 "--account=plgsubslearnath-gpu-a100",
                 "--job-name=lr_grid",
                 "--time=40:00:00",
-                "lizrd/scripts/grid_entrypoint.sh",
+                "lizrd/grid/grid_entrypoint.sh",
                 "singularity",
                 "run",
                 "--bind=/net:/net",
