+++ research/conditional/tests/test_integrated.py
@@ -17,6 +17,6 @@ class TestIntegrated(GeneralTestCase):
         for path in configs:
             print(f"Running training loop with config from {str(path)}")
             exit_code = subprocess.call(
-                ["python3", "-m", "lizrd.scripts.grid", f"--config_path={str(path)}"]
+                ["python3", "-m", "lizrd.grid.grid", f"--config_path={str(path)}"]
             )
             assert exit_code == 0
